fsp_bi_hub/requirements.txt
alembic==1.6.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
google-api-core==1.27.0
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.0
gspread==3.7.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
protobuf==3.17.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.1
